Rename beers.models.models beers.models.contest_player
